@@ -8,8 +8,8 @@ var PartStream = require('./PartStream'),
     HeaderParser = require('./HeaderParser');
 
 var DASH = 45,
-    B_ONEDASH = new Buffer('-'),
-    B_CRLF = new Buffer('\r\n'),
+    B_ONEDASH = Buffer.from('-'),
+    B_CRLF = Buffer.from('\r\n'),
     EMPTY_FN = function() {};
 
 function Dicer(cfg) {
