@@ -3,7 +3,7 @@ var EventEmitter = require('events').EventEmitter,
 
 var StreamSearch = require('streamsearch');
 
-var B_DCRLF = new Buffer('\r\n\r\n'),
+var B_DCRLF = Buffer.from('\r\n\r\n'),
     RE_CRLF = /\r\n/g,
     RE_HDR = /^([^:]+):[ \t]?([\x00-\xFF]+)?$/,
     MAX_HEADER_PAIRS = 2000, // from node's http.js
