@@ -4,7 +4,7 @@ var assert = require('assert'),
 var HeaderParser = require('../lib/HeaderParser');
 
 var DCRLF = '\r\n\r\n',
-    MAXED_BUFFER = new Buffer(128 * 1024);
+    MAXED_BUFFER = Buffer.allocUnsafe(128 * 1024);
 MAXED_BUFFER.fill(0x41); // 'A'
 
 var group = path.basename(__filename, '.js') + '/';
