@@ -58,7 +58,7 @@ function next() {
   var v = tests[t],
       fixtureBase = FIXTURES_ROOT + v.source,
       n = 0,
-      buffer = new Buffer(v.chsize),
+      buffer = Buffer.allocUnsafe(v.chsize),
       state = { parts: [], preamble: undefined };
 
   var dicer = new Dicer(v.opts),
@@ -80,7 +80,7 @@ function next() {
         dicer.setBoundary(v.setBoundary);
     }).on('data', function(data) {
       // make a copy because we are using readSync which re-uses a buffer ...
-      var copy = new Buffer(data.length);
+      var copy = Buffer.allocUnsafe(data.length);
       data.copy(copy);
       data = copy;
       if (!preamble.body)
